修复问题
  - 严重: Admin 常数时间比较 u8 截断（安全绕过）、重复位置跨批次分割（数据丢失）
  - 高: binlog 位置 u32 截断（>4GiB 溢出）、连接超时泄漏复制任务
  - 中: 事件计数器双倍、未使用的 charset/idle_timeout 配置
  - 低: sink 不必要克隆、instances_active gauge 从未赋值

+++ crates/canal-admin/src/lib.rs
@@ -129,11 +129,11 @@ fn constant_time_eq(a: &[u8], b: &[u8]) -> bool {
     // Constant-time comparison — no early return on length mismatch,
     // always runs the full XOR fold to avoid timing side-channel
     let max_len = a.len().max(b.len());
-    let mut acc: u8 = (a.len() ^ b.len()) as u8;
+    let mut acc: u16 = (a.len() ^ b.len()) as u16;
     for i in 0..max_len {
         let x = a.get(i).copied().unwrap_or(0);
         let y = b.get(i).copied().unwrap_or(0);
-        acc |= x ^ y;
+        acc |= (x ^ y) as u16;
     }
     acc == 0
 }
